+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2024-05-14
+### Added
+- Ch.4 code examples for n x n matrix multiplication (simple, recursive and strassen's algorithm) and simple n x n matrix addition
+
 ## 2024-05-11
 ### Fixed
 - Optimised merge sort, merge sort with inversion count and bubble sort